+++ CHANGES.rst
@@ -9,6 +9,7 @@ Features:
 
 Improvements:
  - Remove double negations from settings and update descriptions (#2723)
+ - Handle missing or bad parameter in slash command
 
 Other changes:
  -